+++ index.js
@@ -1,22 +1,19 @@
-var fis = module.exports = require('fis3');
+var ybruin = module.exports = require('fis3');
 
-fis.cli.name = 'ybruin';
-fis.cli.info = require('./package.json');
-fis.require.prefixes.unshift('ybruin', 'fis3');
-fis.cli.version = require('./lib/version.js');
-fis.cli.help = require('./lib/help.js');
-fis.cli.run = require('./lib/run.js');
-fis.file = require('./lib/file.js');
+ybruin.cli.name = 'ybruin';
+ybruin.cli.info = require('./package.json');
+ybruin.require.prefixes.unshift('ybruin', 'ybruin3');
+ybruin.cli.version = require('./lib/version.js');
+ybruin.cli.help = require('./lib/help.js');
+ybruin.cli.run = require('./lib/run.js');
+ybruin.file = require('./lib/file.js');
 
-fis.config.set('project.ignore',['node_modules/**', 'output/**', '.git/**', 'ybruin-conf.js'])
+ybruin.config.set('project.ignore',['node_modules/**', 'output/**', '.git/**', 'ybruin-conf.js'])
 
-//mount scene
-fis.runConf = require('./configs/default.js');
-
-fis.runComponentConf = require('./configs/componentDefault.js');
+yrbuin.runComponentConf = require('./configs/componentDefault.js');
 //alias
 Object.defineProperty(global, 'ybruin', {
   enumerable : true,
   writable : false,
-  value : fis
+  value : ybruin
 });